requirements.txt
google-genai
Pillow
PyMuPDF
flask
python-dotenv
